refactor(ivc): extract mod
